% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylorugD1-package.R
\docType{package}
\name{phylorugD1-package}
\alias{phylorugD1}
\alias{phylorugD1-package}
\title{phylorugD1: Visualize Node Support Across Multiple Phylogenomic Analyses}
\description{
Provides functions to visualise node support values from multiple independent phylogenomic analyses simultaneously on a single reference tree using the rug plot method. Although researchers routinely run multiple phylogenomic analyses, they publish one tree and hide the rest in supplementary materials, yet no dedicated R tool exists to visualise node support across multiple analyses, leaving researchers unable to communicate phylogenomic uncertainty transparently in a single publication figure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdrifathahamed/phylorugD1}
  \item \url{https://mdrifathahamed.github.io/phylorugD1/}
  \item Report bugs at \url{https://github.com/mdrifathahamed/phylorugD1/issues}
}

}
\author{
\strong{Maintainer}: Md Rifath Ahamed \email{rifath.ahamed@helsinki.fi} (\href{https://orcid.org/0009-0005-6158-8658}{ORCID})

Authors:
\itemize{
  \item Md Rifath Ahamed \email{rifath.ahamed@helsinki.fi} (\href{https://orcid.org/0009-0005-6158-8658}{ORCID})
  \item Sergei Tarasov \email{sergei.tarasov@helsinki.fi} (\href{https://orcid.org/0000-0002-1737-9403}{ORCID})
}

}
\keyword{internal}
